+++ CHANGELOG.md
@@ -61,6 +61,7 @@ FlatLaf Change Log
 - Slider: Support specifying width of thumb border (see UI value
   `Slider.thumbBorderWidth`).
 - TabbedPane: Optionally paint selected tab as card. (PR #343)
+- MenuItem: Paint the selected icon when the item is selected. (PR #415)
 - Added more color functions to class `ColorFunctions` for easy use in
   applications: `lighten()`, `darken()`, `saturate()`, `desaturate()`, `spin()`,
   `tint()`, `shade()` and `luma()`.